usbuart: Moved usbuart_set_led_state() into aux_serial.c
